Monitoring_System/requirements.txt
# # conda install pytorch torchvision -c pytorch
# # jsonschema 
# # sniffio 
# # portalocker 
# # scikit-learn 
# # scipy regex

# # --------------------- detector ---------------------------
# matplotlib
# numpy
# opencv-python
# Pillow
# PyYAML
# scipy
# torch
# torchvision
# tqdm

# tensorboard
# wandb

# seaborn
# pandas
# matplotlib

# pycocotools
# thop

# # ------------------------ tracker ----------------------------
# deep-sort-realtime
# portalocker


# # --------------------- pose estimation ---------------------------
# # build.txt
# numpy
# torch>=1.8

# # runtime
# chumpy
# json_tricks
# matplotlib
# munkres
# numpy
# opencv-python
# pillow
# scipy
# torchvision
# xtcocotools>=1.12

# # test
# coverage
# flake8
# interrogate
# isort==4.3.21
# parameterized
# pytest
# pytest-runner
# xdoctest>=0.10.0
# yapf

# # optional
# requests
# facenet-pytorch
# albumentations
# tqdm
# ultralytics


absl-py==2.2.2
addict==2.4.0
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
chumpy==0.70
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
coverage==7.8.0
crcmod==1.7
cryptography==45.0.2
cycler==0.12.1
cython==3.1.0
deep-sort-realtime==1.3.2
dlib==19.24.9
docker-pycreds==0.4.0
eval-type-backport==0.2.2
facenet-pytorch==2.6.0
filelock==3.14.0
flake8==7.2.0
fonttools==4.58.0
grpcio==1.71.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iniconfig==2.1.0
interrogate==1.7.0
isort==4.3.21
jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.0
json-tricks==3.17.3
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
markdown==3.8
markupsafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
mmcv==2.1.0
mmdet==3.3.0
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
munkres==1.1.4
networkx==3.2.1
numpy==1.26.4
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pillow==10.2.0
platformdirs==4.3.8
pluggy==1.6.0
portalocker==3.1.1
protobuf==6.31.0
psutil==7.0.0
py==1.11.0
py-cpuinfo==9.0.0
pycodestyle==2.13.0
pycparser==2.22
pycryptodome==3.23.0
pyflakes==3.3.2
pylance==0.28.0
pyparsing==3.2.3
pytest==8.3.5
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
pytz==2023.4
pywin32==310
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.28.2
rich==13.4.2
rpds-py==0.25.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==60.2.0
shapely==2.0.7
simsimd==6.4.9
six==1.17.0
sniffio==1.3.1
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
terminaltables==3.1.10
thop==0.1.1-2209072238
threadpoolctl==3.6.0
tomli==2.2.1
tqdm==4.65.2
ultralytics==8.3.157
ultralytics-thop==2.0.14
urllib3==1.26.20
wandb==0.19.11
werkzeug==3.1.3
xdoctest==1.2.0
xtcocotools==1.14.3
yapf==0.43.0
zipp==3.21.0